🚨WIP: Move pre-release filtering into DependencyCache
I haven't cleaned up the Repository code for this change, and the tests
that test for this behaviour need to be (I guess?) changed to use
DependencyCache instead.

I also didn't limit the fallback-to-prerelease behaviour to the "*"
constraint, since that will still break for the second new test case, I
think.

So this probably isn't the right solution, as we still have the problem
that we abort solving if DependencyCache returns no hits, but continue
solving if DepdendencyCache returns a set of hits, and a later
constraint eliminates _all_ of those hits, and hence exposes an
acceptible pre-release.

So perhaps "don't take pre-releases if a matching release exists" should
be done later in the solver flow, once the total aggregate constraint on
a package has been determined?